requirements.txt
Django
markdown
uwsgi
graphene_django
django-colorfield